Tweaks as per discussion, also upped page size from 50 to 100 for permissions fetch
